ai-service/requirements.txt
fastapi
uvicorn
requests
python-dotenv
openai
groq
pymongo
faiss-cpu
sentence-transformers
numpy
